refactoring

+++ crates/perroute-command-bus/src/commands/business_unit/delete.rs
@@ -3,11 +3,17 @@ use crate::bus::{
 };
 use bon::Builder;
 use perroute_commons::{commands::CommandType, events::Event, types::id::Id};
-use perroute_storage::repository::{business_units::BusinessUnitRepository, TransactedRepository};
+use perroute_storage::repository::{
+    business_units::{BusinessUnitQuery, BusinessUnitRepository},
+    TransactedRepository,
+};
 use serde::Serialize;
 
 #[derive(Debug, thiserror::Error)]
-pub enum DeleteBusinessUnitCommandError {}
+pub enum DeleteBusinessUnitCommandError {
+    #[error("Business unit not found")]
+    BusinessUnitNotFound,
+}
 
 #[derive(Debug, Clone, Builder, Serialize)]
 pub struct DeleteBusinessUnitCommand {
@@ -24,17 +30,26 @@ pub struct DeleteBusinessUnitCommandHandler;
 
 impl CommandHandler for DeleteBusinessUnitCommandHandler {
     type Command = DeleteBusinessUnitCommand;
-    type Output = bool;
+    type Output = ();
 
     async fn handle<R: TransactedRepository>(
         &self,
         cmd: &Self::Command,
         ctx: CommandBusContext<'_, R>,
     ) -> CommandHandlerResult<Self::Output> {
-        let deleted =
-            BusinessUnitRepository::delete_business_unit(ctx.repository(), &cmd.id).await?;
+        let exists = BusinessUnitRepository::exists_business_unit(
+            ctx.repository(),
+            &BusinessUnitQuery::ById(cmd.id.clone()),
+        )
+        .await?;
 
-        CommandHandlerOutput::new(deleted)
+        if !exists {
+            return Err(DeleteBusinessUnitCommandError::BusinessUnitNotFound.into());
+        }
+
+        BusinessUnitRepository::delete_business_unit(ctx.repository(), &cmd.id).await?;
+
+        CommandHandlerOutput::new(())
             .with_event(Event::BusinessUnitDeleted(cmd.id.clone()))
             .ok()
     }
